[MOV] format: move format helpers inside a dedicated folder
Task: 4083126
